requirements.txt
gradio
faiss-cpu
sentence-transformers
scikit-learn
pandas
numpy
nltk
